+++ packages/entities/entities-certificates/sandbox/pages/CertificateFormPage.vue
@@ -41,6 +41,7 @@ const konnectConfig = ref<KonnectCertificateFormConfig>({
   // Set the root `.env.development.local` variable to a control plane your PAT can access
   controlPlaneId,
   cancelRoute: { name: 'certificate-list' },
+  jsonYamlFormsEnabled: true,
 })
 
 const kongManagerConfig = ref<KongManagerCertificateFormConfig>({