+++ packages/entities/entities-certificates/src/components/CertificateForm.vue
@@ -6,7 +6,7 @@
       :edit-id="certificateId"
       :error-message="form.errorMessage"
       :fetch-url="fetchUrl"
-      :form-fields="form.fields"
+      :form-fields="requestBody"
       :is-readonly="form.isReadonly"
       @cancel="handleClickCancel"
       @fetch:error="(err: any) => $emit('error', err)"
@@ -255,30 +255,32 @@ const submitUrl = computed<string>(() => {
   return url
 })
 
+const requestBody = computed((): Record<string, any> => {
+  return {
+    cert: form.fields.cert,
+    key: form.fields.key,
+    cert_alt: form.fields.certAlt || null,
+    key_alt: form.fields.keyAlt || null,
+    tags: form.fields.tags.split(',')?.map((tag: string) => String(tag || '').trim())?.filter((tag: string) => tag !== ''),
+  }
+})
+
 const saveFormData = async (): Promise<void> => {
   try {
     form.isReadonly = true
 
-    const requestBody: Record<string, any> = {
-      cert: form.fields.cert,
-      key: form.fields.key,
-      cert_alt: form.fields.certAlt || null,
-      key_alt: form.fields.keyAlt || null,
-      tags: form.fields.tags.split(',')?.map((tag: string) => String(tag || '').trim())?.filter((tag: string) => tag !== ''),
-    }
-
     let response: AxiosResponse | undefined
 
-    await axiosInstance.post(validateSubmitUrl.value, requestBody)
+    await axiosInstance.post(validateSubmitUrl.value, requestBody.value)
 
     if (formType.value === 'create') {
-      response = await axiosInstance.post(submitUrl.value, requestBody)
+      response = await axiosInstance.post(submitUrl.value, requestBody.value)
     } else if (formType.value === 'edit') {
       response = props.config?.app === 'konnect'
         // Note: Konnect currently uses PUT because PATCH is not fully supported in Koko
         //       If this changes, the `edit` form methods should be re-evaluated/updated accordingly
-        ? await axiosInstance.put(submitUrl.value, requestBody)
-        : await axiosInstance.patch(submitUrl.value, requestBody)
+        ? await axiosInstance.put(submitUrl.value, requestBody.value)
+        : await axiosInstance.patch(submitUrl.value, requestBody.value)
     }
 
     form.fields.cert = response?.data?.cert || ''