+++ packages/entities/entities-key-sets/sandbox/pages/KeySetFormPage.vue
@@ -38,6 +38,7 @@ const konnectConfig = ref<KonnectKeySetFormConfig>({
   // Set the root `.env.development.local` variable to a control plane your PAT can access
   controlPlaneId,
   cancelRoute: { name: 'key-set-list' },
+  jsonYamlFormsEnabled: true,
 })
 const kongManagerConfig = ref<KongManagerKeySetFormConfig>({
   app: 'kongManager',