+++ packages/entities/entities-routes/sandbox/pages/RouteFormPage.vue
@@ -67,6 +67,7 @@ const konnectConfig = ref<KonnectRouteFormConfig>({
   // Set the root `.env.development.local` variable to a control plane your PAT can access
   controlPlaneId,
   cancelRoute: { name: 'route-list' },
+  jsonYamlFormsEnabled: true,
 })
 
 const kongManagerConfig = ref<KongManagerRouteFormConfig>({