+++ packages/entities/entities-snis/sandbox/pages/SniFormPage.vue
@@ -41,8 +41,10 @@ const konnectConfig = ref<KonnectSniFormConfig>({
   // Set the root `.env.development.local` variable to a control plane your PAT can access
   controlPlaneId,
   cancelRoute: { name: 'snis-list' },
+  jsonYamlFormsEnabled: true,
   // uncomment to see create with provided certificate
   /* certificateId: '1234-ilove-cats', */
+
 })
 
 const kongManagerConfig = ref<KongManagerSniFormConfig>({